Title: 일렬로 서있는 소 2, Time: 143ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 260217/일렬로 서있는 소 2/cattle-in-a-rowing-up-2.java
@@ -9,7 +9,7 @@ public class Main {
             arr[i] = sc.nextInt();
         // Please write your code here.
 
-        int cnt = 0;
+        long cnt = 0;
 
         for (int i=0; i<n; i++) {
             for (int j=i+1; j<n; j++) {
